Extract BattleEvalConstants & ValueRecorder into full class ; move VR into util ; add CreatureAttackTacticalObjective (unused yet), which duplicates a lot of code from SimpleAI ATM ; priority are float now (to avoid excess in CreatureAttackTacticalObjective, we can use priority < 1.f) ; drop idiotic countCreatureType (Legion has numCreature...).
git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@4261 c5679a2a-0e0e-0410-bba2-8642b794b8c1

+++ Colossus/core/src/main/java/net/sf/colossus/ai/AbstractTacticalObjective.java
@@ -1,25 +1,26 @@
 package net.sf.colossus.ai;
 
-/**
+/** Abstract implementation of TacticalObjective, handling the priority
+ * stuff to avoid duplication.
  *
  * @author Romain Dolbeau
  */
 public abstract class AbstractTacticalObjective implements TacticalObjective {
-    private int priority;
+    private float priority;
 
-    public AbstractTacticalObjective(int priority)
+    public AbstractTacticalObjective(float priority)
     {
         this.priority = priority;
     }
 
-    public int getPriority()
+    public float getPriority()
     {
         return priority;
     }
 
-    public int changePriority(int newPriority)
+    public float changePriority(float newPriority)
     {
-        int oldPriority = priority;
+        float oldPriority = priority;
         priority = newPriority;
         return oldPriority;
     }
